+++ RayTracer.MathLibrary.Tests/Point3DTests.cs
@@ -10,9 +10,9 @@ namespace RayTracer.MathLibrary.Tests
         [Test]
         [TestCase(1f, -1f, 0f, 0f, 0f, 1f, 1f, -1f, -1f)]
         [TestCase(1f, 1f, 6f, 3f, -2f, 5f, -2f, 3f, 1f)]
-        public void AdditionOperator_WhenCalledWithTwoPoints_ReturnPointWithCombinedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                               float arg1X, float arg1Y, float arg1Z,
-                                                                                               float arg2X, float arg2Y, float arg2Z)
+        public void AdditionOperator_WhenCalledWithTwoPoints_ReturnPointWithCombinedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                               double arg1X, double arg1Y, double arg1Z,
+                                                                                               double arg2X, double arg2Y, double arg2Z)
         {
             Point3D point1 = new Point3D()
             {
@@ -41,9 +41,9 @@ namespace RayTracer.MathLibrary.Tests
         [Test]
         [TestCase(1f, -1f, 0f, 0f, 0f, 1f, 1f, -1f, -1f)]
         [TestCase(1f, 1f, 6f, 3f, -2f, 5f, -2f, 3f, 1f)]
-        public void AdditionOperator_WhenCalledWithPointAndVector_ReturnPointWithCombinedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                                    float arg1X, float arg1Y, float arg1Z,
-                                                                                                    float arg2X, float arg2Y, float arg2Z)
+        public void AdditionOperator_WhenCalledWithPointAndVector_ReturnPointWithCombinedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                                    double arg1X, double arg1Y, double arg1Z,
+                                                                                                    double arg2X, double arg2Y, double arg2Z)
         {
             Point3D point = new Point3D()
             {
@@ -72,9 +72,9 @@ namespace RayTracer.MathLibrary.Tests
         [Test]
         [TestCase(-1f, 1f, 2f, 0f, 0f, 1f, 1f, -1f, -1f)]
         [TestCase(-2f, -4f, -6f, 3f, 2f, 1f, 5f, 6f, 7f)]
-        public void SubtractionOperator_WhenCalledWithTwoPoints_ReturnPointWithSubtractedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                                    float arg1X, float arg1Y, float arg1Z,
-                                                                                                    float arg2X, float arg2Y, float arg2Z)
+        public void SubtractionOperator_WhenCalledWithTwoPoints_ReturnPointWithSubtractedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                                    double arg1X, double arg1Y, double arg1Z,
+                                                                                                    double arg2X, double arg2Y, double arg2Z)
         {
             Point3D point1 = new Point3D()
             {
@@ -103,9 +103,9 @@ namespace RayTracer.MathLibrary.Tests
         [Test]
         [TestCase(-1f, 1f, 2f, 0f, 0f, 1f, 1f, -1f, -1f)]
         [TestCase(-2f, -4f, -6f, 3f, 2f, 1f, 5f, 6f, 7f)]
-        public void SubtractionOperator_WhenCalledWithPointAndVector_ReturnPointWithSubtractedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                                         float arg1X, float arg1Y, float arg1Z,
-                                                                                                         float arg2X, float arg2Y, float arg2Z)
+        public void SubtractionOperator_WhenCalledWithPointAndVector_ReturnPointWithSubtractedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                                         double arg1X, double arg1Y, double arg1Z,
+                                                                                                         double arg2X, double arg2Y, double arg2Z)
         {
             Point3D point = new Point3D()
             {
@@ -136,8 +136,8 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(3f, 3f, 3f, 2f, 2f, 2f, 1.5f)]
         [TestCase(3.5f, -7f, 10.5f, 1f, -2f, 3f, 3.5f)]
         [TestCase(0.5f, -1f, 1.5f, 1f, -2f, 3f, 0.5f)]
-        public void MultiplyOperator_WhenCalledWithPointAndScalar_ReturnMultipliedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                             float argX, float argY, float argZ, float scalar)
+        public void MultiplyOperator_WhenCalledWithPointAndScalar_ReturnMultipliedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                             double argX, double argY, double argZ, double scalar)
         {
             Point3D expected = new Point3D()
             {
@@ -160,8 +160,8 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(1f, 1f, 1f, 2f, 2f, 2f, 2f)]
         [TestCase(2f, 2f, 2f, 3f, 3f, 3f, 1.5f)]
         [TestCase(0.5f, -1f, 1.5f, 1f, -2f, 3f, 2f)]
-        public void DivideOperator_WhenCalledWithPointAndScalar_ReturnMultipliedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                           float argX, float argY, float argZ, float scalar)
+        public void DivideOperator_WhenCalledWithPointAndScalar_ReturnMultipliedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                           double argX, double argY, double argZ, double scalar)
         {
             Point3D expected = new Point3D()
             {
